+++ BALSAMIC/commands/config/pon.py
@@ -9,19 +9,21 @@ import click
 
 from BALSAMIC.commands.options import (
     OPTION_GENOME_VERSION,
+    OPTION_GENOME_INTERVAL,
     OPTION_ADAPTER_TRIM,
     OPTION_ANALYSIS_DIR,
     OPTION_BALSAMIC_CACHE,
     OPTION_CASE_ID,
     OPTION_FASTQ_PATH,
     OPTION_PANEL_BED,
+    OPTION_PON_WORKFLOW,
     OPTION_PON_VERSION,
     OPTION_QUALITY_TRIM,
     OPTION_UMI,
     OPTION_UMI_TRIM_LENGTH,
     OPTION_CACHE_VERSION,
 )
-from BALSAMIC.constants.analysis import BIOINFO_TOOL_ENV
+from BALSAMIC.constants.analysis import BIOINFO_TOOL_ENV, PONWorkflow
 from BALSAMIC.constants.cache import GenomeVersion
 from BALSAMIC.constants.constants import FileType
 from BALSAMIC.constants.paths import CONTAINERS_DIR
@@ -46,7 +48,9 @@ LOG = logging.getLogger(__name__)
 @OPTION_CASE_ID
 @OPTION_FASTQ_PATH
 @OPTION_GENOME_VERSION
+@OPTION_GENOME_INTERVAL
 @OPTION_PANEL_BED
+@OPTION_PON_WORKFLOW
 @OPTION_PON_VERSION
 @OPTION_QUALITY_TRIM
 @OPTION_UMI
@@ -61,7 +65,9 @@ def pon_config(
     case_id: str,
     fastq_path: Path,
     genome_version: GenomeVersion,
+    genome_interval: Path,
     panel_bed: Path,
+    pon_workflow: PONWorkflow,
     quality_trim: bool,
     umi: bool,
     umi_trim_length: bool,
@@ -72,6 +78,19 @@ def pon_config(
         base_path=references_path,
         data=read_json(Path(references_path, f"reference.{FileType.JSON}").as_posix()),
     )
+
+    if pon_workflow in [PONWorkflow.GENS_MALE, PONWorkflow.GENS_FEMALE]:
+        if not genome_interval:
+            raise click.BadParameter(
+                "Argument: genome_interval is required for GENS PON creation."
+            )
+        references["genome_interval"] = genome_interval
+
+    if pon_workflow == PONWorkflow.CNVKIT and not panel_bed:
+        raise click.BadParameter(
+            "Argument: panel_bed is required for CNVkit PON creation."
+        )
+
     fastq_path: str = get_analysis_fastq_files_directory(
         case_dir=Path(analysis_dir, case_id).as_posix(), fastq_path=fastq_path
     )
@@ -87,6 +106,7 @@ def pon_config(
             "analysis_dir": analysis_dir,
             "fastq_path": fastq_path,
             "analysis_type": "pon",
+            "pon_workflow": pon_workflow,
             "pon_version": version,
             "analysis_workflow": "balsamic",
             "sequencing_type": "targeted" if panel_bed else "wgs",